+++ app/lib/core/native_adapter_instance.dart
